fix: rename member-types folder
